feat(core): implement wxFor and refactor related code
