+++ packages/language-core/src/types/compiler.ts
@@ -1,0 +1,8 @@
+import * as CompilerDOM from '@vue/compiler-dom'
+
+export type Node =
+  | CompilerDOM.RootNode
+  | CompilerDOM.TemplateChildNode
+  | CompilerDOM.ExpressionNode
+  | CompilerDOM.AttributeNode
+  | CompilerDOM.DirectiveNode
